rename `logging` to `logging_utils` trying to fix win-py37 issue on CI
